requirements.txt
beautifulsoup4
groq
python-dotenv
requests
streamlit
